requirements.txt
 pandas
 selenium
 webdriver_manager>=4.0.0
 Pillow
 openai==0.28
 pytesseract
 python-dotenv
 anthropic
 requests
requests
beautifulsoup4
